[orm]: Fix update one-and-only pg assertion
