+++ db/index.ts
@@ -1,3 +1,4 @@
+export * from './orm'
 export * from './fpprg'
 export * from './accounts'
 export * from './suggestions'
